Added everything from tutorial #7

+++ src/com/gametutorial/main/BasicEnemy.java
@@ -5,7 +5,7 @@ import java.awt.*;
 public class BasicEnemy extends GameObject{
     private Handler handler;
 
-    public BasicEnemy(int x, int y, ID id, Handler handler) {
+    public BasicEnemy(float x, float y, ID id, Handler handler) {
         super(x, y, id);
 
         this.handler = handler;
@@ -32,11 +32,11 @@ public class BasicEnemy extends GameObject{
     @Override
     public void render(Graphics g) {
         g.setColor(Color.red);
-        g.fillRect(x, y, 16, 16);
+        g.fillRect((int)x, (int)y, 16, 16);
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 16, 16);
+        return new Rectangle((int)x, (int)y, 16, 16);
     }
 }
